Revert "refactor: split hooks"
