--- src/pages/Rekap/Rekap.hooks.jsx
@@ -1,16 +1,0 @@
-import {useEffect} from 'react'
-import { useNavigate } from 'react-router-dom'
-import { useSelector } from 'react-redux'
-
-const useRekap = () => {
-    const navigate = useNavigate()
-    const { user } = useSelector((state) => state.login)
-
-    useEffect(() => {
-        if(user?.admin === false) {
-            navigate('/')
-        }
-    }, [])
-}
-
-export default useRekap
