+++ src/pages/Rekap/index.jsx
@@ -1,9 +1,18 @@
-import useRekap from './Rekap.hooks'
+import React, {useEffect} from 'react'
+import { useNavigate } from 'react-router-dom'
+import { useSelector } from 'react-redux'
 import RekapTable from '../../components/organisms/RekapTable'
 
 const Rekap = () => {
-  useRekap()
-  
+  const navigate = useNavigate()
+  const { user } = useSelector((state) => state.login)
+
+  useEffect(() => {
+    if(user?.admin === false) {
+      navigate('/')
+    }
+  }, [])
+
   return (
     <RekapTable />
   )
